requirements.txt
Django
Pillow
whitenoise
django_neomodel
mssql-django
django-cors-headers